+++ lib/src/theme/text_theme.freezed.dart
@@ -1,0 +1,134 @@
+// dart format width=80
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'text_theme.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+// dart format off
+T _$identity<T>(T value) => value;
+/// @nodoc
+mixin _$WaveTextTheme {
+
+ TextStyle get body;
+/// Create a copy of WaveTextTheme
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$WaveTextThemeCopyWith<WaveTextTheme> get copyWith => _$WaveTextThemeCopyWithImpl<WaveTextTheme>(this as WaveTextTheme, _$identity);
+
+
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is WaveTextTheme&&(identical(other.body, body) || other.body == body));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,body);
+
+
+
+}
+
+/// @nodoc
+abstract mixin class $WaveTextThemeCopyWith<$Res>  {
+  factory $WaveTextThemeCopyWith(WaveTextTheme value, $Res Function(WaveTextTheme) _then) = _$WaveTextThemeCopyWithImpl;
+@useResult
+$Res call({
+ TextStyle body
+});
+
+
+
+
+}
+/// @nodoc
+class _$WaveTextThemeCopyWithImpl<$Res>
+    implements $WaveTextThemeCopyWith<$Res> {
+  _$WaveTextThemeCopyWithImpl(this._self, this._then);
+
+  final WaveTextTheme _self;
+  final $Res Function(WaveTextTheme) _then;
+
+/// Create a copy of WaveTextTheme
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? body = null,}) {
+  return _then(_self.copyWith(
+body: null == body ? _self.body : body // ignore: cast_nullable_to_non_nullable
+as TextStyle,
+  ));
+}
+
+}
+
+
+/// @nodoc
+
+
+class _WaveTextTheme extends WaveTextTheme {
+  const _WaveTextTheme({required this.body}): super._();
+  
+
+@override final  TextStyle body;
+
+/// Create a copy of WaveTextTheme
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$WaveTextThemeCopyWith<_WaveTextTheme> get copyWith => __$WaveTextThemeCopyWithImpl<_WaveTextTheme>(this, _$identity);
+
+
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _WaveTextTheme&&(identical(other.body, body) || other.body == body));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,body);
+
+
+
+}
+
+/// @nodoc
+abstract mixin class _$WaveTextThemeCopyWith<$Res> implements $WaveTextThemeCopyWith<$Res> {
+  factory _$WaveTextThemeCopyWith(_WaveTextTheme value, $Res Function(_WaveTextTheme) _then) = __$WaveTextThemeCopyWithImpl;
+@override @useResult
+$Res call({
+ TextStyle body
+});
+
+
+
+
+}
+/// @nodoc
+class __$WaveTextThemeCopyWithImpl<$Res>
+    implements _$WaveTextThemeCopyWith<$Res> {
+  __$WaveTextThemeCopyWithImpl(this._self, this._then);
+
+  final _WaveTextTheme _self;
+  final $Res Function(_WaveTextTheme) _then;
+
+/// Create a copy of WaveTextTheme
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? body = null,}) {
+  return _then(_WaveTextTheme(
+body: null == body ? _self.body : body // ignore: cast_nullable_to_non_nullable
+as TextStyle,
+  ));
+}
+
+
+}
+
+// dart format on