+++ lib/src/widgets/app_bar/app_bar_theme.freezed.dart
@@ -1,0 +1,143 @@
+// dart format width=80
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'app_bar_theme.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+// dart format off
+T _$identity<T>(T value) => value;
+/// @nodoc
+mixin _$WaveAppBarTheme {
+
+ Color get backgroundColor; Color get foregroundColor; TextStyle get titleStyle; bool get isCenteredTitle;
+/// Create a copy of WaveAppBarTheme
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$WaveAppBarThemeCopyWith<WaveAppBarTheme> get copyWith => _$WaveAppBarThemeCopyWithImpl<WaveAppBarTheme>(this as WaveAppBarTheme, _$identity);
+
+
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is WaveAppBarTheme&&(identical(other.backgroundColor, backgroundColor) || other.backgroundColor == backgroundColor)&&(identical(other.foregroundColor, foregroundColor) || other.foregroundColor == foregroundColor)&&(identical(other.titleStyle, titleStyle) || other.titleStyle == titleStyle)&&(identical(other.isCenteredTitle, isCenteredTitle) || other.isCenteredTitle == isCenteredTitle));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,backgroundColor,foregroundColor,titleStyle,isCenteredTitle);
+
+
+
+}
+
+/// @nodoc
+abstract mixin class $WaveAppBarThemeCopyWith<$Res>  {
+  factory $WaveAppBarThemeCopyWith(WaveAppBarTheme value, $Res Function(WaveAppBarTheme) _then) = _$WaveAppBarThemeCopyWithImpl;
+@useResult
+$Res call({
+ Color backgroundColor, Color foregroundColor, TextStyle titleStyle, bool isCenteredTitle
+});
+
+
+
+
+}
+/// @nodoc
+class _$WaveAppBarThemeCopyWithImpl<$Res>
+    implements $WaveAppBarThemeCopyWith<$Res> {
+  _$WaveAppBarThemeCopyWithImpl(this._self, this._then);
+
+  final WaveAppBarTheme _self;
+  final $Res Function(WaveAppBarTheme) _then;
+
+/// Create a copy of WaveAppBarTheme
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? backgroundColor = null,Object? foregroundColor = null,Object? titleStyle = null,Object? isCenteredTitle = null,}) {
+  return _then(_self.copyWith(
+backgroundColor: null == backgroundColor ? _self.backgroundColor : backgroundColor // ignore: cast_nullable_to_non_nullable
+as Color,foregroundColor: null == foregroundColor ? _self.foregroundColor : foregroundColor // ignore: cast_nullable_to_non_nullable
+as Color,titleStyle: null == titleStyle ? _self.titleStyle : titleStyle // ignore: cast_nullable_to_non_nullable
+as TextStyle,isCenteredTitle: null == isCenteredTitle ? _self.isCenteredTitle : isCenteredTitle // ignore: cast_nullable_to_non_nullable
+as bool,
+  ));
+}
+
+}
+
+
+/// @nodoc
+
+
+class _WaveAppBarTheme extends WaveAppBarTheme {
+  const _WaveAppBarTheme({required this.backgroundColor, required this.foregroundColor, required this.titleStyle, this.isCenteredTitle = true}): super._();
+  
+
+@override final  Color backgroundColor;
+@override final  Color foregroundColor;
+@override final  TextStyle titleStyle;
+@override@JsonKey() final  bool isCenteredTitle;
+
+/// Create a copy of WaveAppBarTheme
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$WaveAppBarThemeCopyWith<_WaveAppBarTheme> get copyWith => __$WaveAppBarThemeCopyWithImpl<_WaveAppBarTheme>(this, _$identity);
+
+
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _WaveAppBarTheme&&(identical(other.backgroundColor, backgroundColor) || other.backgroundColor == backgroundColor)&&(identical(other.foregroundColor, foregroundColor) || other.foregroundColor == foregroundColor)&&(identical(other.titleStyle, titleStyle) || other.titleStyle == titleStyle)&&(identical(other.isCenteredTitle, isCenteredTitle) || other.isCenteredTitle == isCenteredTitle));
+}
+
+
+@override
+int get hashCode => Object.hash(runtimeType,backgroundColor,foregroundColor,titleStyle,isCenteredTitle);
+
+
+
+}
+
+/// @nodoc
+abstract mixin class _$WaveAppBarThemeCopyWith<$Res> implements $WaveAppBarThemeCopyWith<$Res> {
+  factory _$WaveAppBarThemeCopyWith(_WaveAppBarTheme value, $Res Function(_WaveAppBarTheme) _then) = __$WaveAppBarThemeCopyWithImpl;
+@override @useResult
+$Res call({
+ Color backgroundColor, Color foregroundColor, TextStyle titleStyle, bool isCenteredTitle
+});
+
+
+
+
+}
+/// @nodoc
+class __$WaveAppBarThemeCopyWithImpl<$Res>
+    implements _$WaveAppBarThemeCopyWith<$Res> {
+  __$WaveAppBarThemeCopyWithImpl(this._self, this._then);
+
+  final _WaveAppBarTheme _self;
+  final $Res Function(_WaveAppBarTheme) _then;
+
+/// Create a copy of WaveAppBarTheme
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? backgroundColor = null,Object? foregroundColor = null,Object? titleStyle = null,Object? isCenteredTitle = null,}) {
+  return _then(_WaveAppBarTheme(
+backgroundColor: null == backgroundColor ? _self.backgroundColor : backgroundColor // ignore: cast_nullable_to_non_nullable
+as Color,foregroundColor: null == foregroundColor ? _self.foregroundColor : foregroundColor // ignore: cast_nullable_to_non_nullable
+as Color,titleStyle: null == titleStyle ? _self.titleStyle : titleStyle // ignore: cast_nullable_to_non_nullable
+as TextStyle,isCenteredTitle: null == isCenteredTitle ? _self.isCenteredTitle : isCenteredTitle // ignore: cast_nullable_to_non_nullable
+as bool,
+  ));
+}
+
+
+}
+
+// dart format on