+++ lib/theme.dart
@@ -2,4 +2,4 @@
 library waveui.theme;
 
 export 'src/theme/color_scheme.dart' show WaveColorScheme;
-export 'src/theme/theme_data.dart' show WaveThemeData;
+export 'src/theme/theme.dart' show WaveTheme;