+++ lib/widgets/app_bar.dart
@@ -1,0 +1,2 @@
+export '../src/widgets/app_bar/app_bar.dart' show WaveAppBar;
+export '../src/widgets/app_bar/app_bar_theme.dart' show WaveAppBarTheme;